+++ src/components/cookifyInput.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useCookifyProvider } from '../context/cookifyContext'
 import { CookifyInputProps } from '../types'
 
-export const CookifyInput: React.FC<CookifyInputProps> = (props) => {
+export const CookifyInput: React.FC<CookifyInputProps> = ({ name, ...rest }) => {
     const {consentObject, actionCheckbox} = useCookifyProvider()
 
     const handleChange: React.ChangeEventHandler<HTMLInputElement> = (e) => {
@@ -11,8 +11,9 @@ export const CookifyInput: React.FC<CookifyInputProps> = (props) => {
 
     return (
         <input
-            {...props}
-            checked={consentObject.data[props.name]}
+            {...rest}
+            name={name}
+            checked={consentObject.data[name]}
             onChange={e => handleChange(e)}
         />
     )