+++ src/components/index.ts
@@ -1,1 +1,2 @@
 export { default as CookifyInput } from './cookifyInput'
+export { default as CookifyConsent } from './consent/cookifyConsent'